example/requirements.txt
asgiref==3.5.0
Django
Pillow
pytz==2019.3
sqlparse==0.3.0

pytest-django==4.5.2
django-storages==1.12.3
google-cloud-storage==2.3.0